:recycle: refactor(books): Move hooks to `/ui`.
To minimize server refreshes when developing.
