+++ imports/ui/books/useBooks.ts
@@ -1,0 +1,7 @@
+import makeQuery from '../../api/makeQuery';
+import publication from '../../api/publication/books/find';
+import {Books} from '../../api/collection/books';
+
+const useBooks = makeQuery(Books, publication);
+
+export default useBooks;
